Add metric for tracking how long a request is waiting in the queue
This works by adding another attribute "received" to the Redis object that
serializes the request, and using that attribute when the request is being
handled by a worker to calculate the time it spent in the queue. This time is
recorded in a Prometheus histogram metric `coco_queue_wait_time`, with a label
for the "endpoint".