feat(state): commands to save and load states
- refactor: simplify coco_runner arguments (for testing)
- Add internal endpoints and client commands to save and load endpoints.
  New commands / internal endpoints:
  - save-state
  - load-state
  - saved-states

  Parts of the state configured to be excluded from resets are also
  excluded from being loaded from a previously loaded state. But they do
  get saved.

BREAKING CHANGE:
The config variable 'storage_path' is not a file anymore, but a
directory. The state is saved there under the name 'active' and saved
states are added here as well. If you are updating your coco instance,
you have to manually move the persistent state from 'storage_path' to
'storage_path'/active.